Self-review fixes
