+++ framework/arcane-framework/src/main/scala/services/lakehouse/base/IcebergCatalogSettings.scala
@@ -35,5 +35,5 @@ trait IcebergCatalogSettings:
   /**
    * The lakehouse location of the catalog
    */
-  val locationOverride: Option[String]
+  val stagingLocation: Option[String]
