+++ plugin/arcane-stream-sqlserver-change-tracking/integration-tests.env
@@ -2,7 +2,7 @@ STREAMCONTEXT__BACKFILL=false
 STREAMCONTEXT__SPEC='{ "database": "IntegrationTests", "schema": "dbo", "table": "TestTable", "changeCaptureIntervalSeconds": 15, "commandTimeout": 3600, "groupingIntervalSeconds": 15, "groupsPerFile": 1, "lookBackInterval": 3600, "rowsPerGroup": 10000, "sinkLocation": "s3a://integration-tests", "icebergNamespace": "test", "icebergWarehouse": "polaris", "icebergCatalogUri": "http://localhost:8181/api/catalog" }'
 STREAMCONTEXT__STREAM_ID=test
 STREAMCONTEXT__STREAM_KIND=SqlServerChangeTracking
-ARCANE.STREAM.SQL_SERVER_CHANGE_TRACKING__ARCANE_CONNECTION_STRING='jdbc:sqlserver://localhost:1433;databaseName=IntegrationTests;user=sa;password=tMIxN11yGZgMC;encrypt=false;trustServerCertificate=true'
+ARCANE_STREAM__SQL_SERVER_CHANGE_TRACKING__ARCANE_CONNECTION_STRING='jdbc:sqlserver://localhost:1433;databaseName=IntegrationTests;user=sa;password=tMIxN11yGZgMC;encrypt=false;trustServerCertificate=true'
 APPLICATION_VERSION=0.0.1
 ARCANE_DATADOG_ENDPOINT=tcp-intake.logs.datadoghq.eu:443
